move socks port option out to separate class
