+++ sampleapp/src/main/java/io/matthewnelson/sampleapp/ui/fragments/settings/tor/SettingsTorFragment.kt
@@ -66,9 +66,7 @@
 * */
 package io.matthewnelson.sampleapp.ui.fragments.settings.tor
 
-import android.content.Context
 import android.os.Bundle
-import android.text.InputFilter
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -79,24 +77,17 @@ import io.matthewnelson.sampleapp.ui.fragments.dashboard.DashMessage
 import io.matthewnelson.sampleapp.ui.fragments.dashboard.DashboardFragment
 import io.matthewnelson.topl_service.TorServiceController
 import io.matthewnelson.topl_service.service.components.onionproxy.ServiceTorSettings
-import io.matthewnelson.topl_core_base.BaseConsts.PortOption
 
 class SettingsTorFragment : Fragment() {
 
-    private companion object {
+    companion object {
         const val AUTO = "Auto"
         const val DISABLED = "Disabled"
         const val CUSTOM = "Custom"
     }
 
     private lateinit var serviceTorSettings: ServiceTorSettings
-
-    // Views
-    private lateinit var spinnerSocksPort: Spinner
-    private lateinit var adapterSocksPort: ArrayAdapter<String>
-    private lateinit var listenerSocksPort: TorOptionSelectionListener
-    private lateinit var textViewSocksPortCustom: TextView
-    private lateinit var editTextSocksPortCustom: EditText
+    private lateinit var socksPortOption: SocksPortOption
     private lateinit var buttonSave: Button
 
     override fun onCreateView(
@@ -110,10 +101,8 @@ class SettingsTorFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        setInitialValues()
+        socksPortOption = SocksPortOption(view, serviceTorSettings)
         findViews(view)
-        setViewParameters()
-        initSpinnerTorSocksPort(view.context)
         setButtonClickListener()
     }
 
@@ -121,146 +110,17 @@ class SettingsTorFragment : Fragment() {
         super.onDestroyView()
     }
 
-    private fun setInitialValues() {
-        initialSocksPort = serviceTorSettings.socksPort
-    }
-
     private fun findViews(view: View) {
-        spinnerSocksPort = view.findViewById(R.id.settings_tor_spinner_socks_port)
-        textViewSocksPortCustom = view.findViewById(R.id.settings_tor_text_view_socks_port_custom)
-        editTextSocksPortCustom = view.findViewById(R.id.settings_tor_edit_text_socks_port_custom)
         buttonSave = view.findViewById(R.id.settings_tor_button_save)
-    }
-
-    private fun setViewParameters() {
-        editTextSocksPortCustom.filters = arrayOf(InputFilter.LengthFilter(5))
     }
 
     private fun setButtonClickListener() {
         buttonSave.setOnClickListener {
-            saveSocksPort() ?: return@setOnClickListener
+            socksPortOption.saveSocksPort() ?: return@setOnClickListener
 
             DashboardFragment.showMessage(
                 DashMessage("Settings Saved", R.drawable.dash_message_color_green, 3_000L)
             )
-        }
-    }
-
-    //////////////////
-    /// Socks Port ///
-    //////////////////
-    private lateinit var initialSocksPort: String
-
-    private fun initSpinnerTorSocksPort(context: Context) {
-        val categorySocksPort = arrayOf(
-            AUTO,
-            CUSTOM,
-            DISABLED
-        )
-        adapterSocksPort = ArrayAdapter(context, R.layout.spinner_list_item, categorySocksPort)
-        adapterSocksPort.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-        spinnerSocksPort.adapter = adapterSocksPort
-        setSocksPortSpinnerValue()
-        listenerSocksPort = TorOptionSelectionListener()
-        spinnerSocksPort.onItemSelectedListener = listenerSocksPort
-    }
-
-    private fun setSocksPortSpinnerValue() {
-        when (initialSocksPort) {
-            PortOption.AUTO -> {
-                socksPortCustomVisibility(false)
-                spinnerSocksPort.setSelection(0)
-            }
-            PortOption.DISABLED -> {
-                socksPortCustomVisibility(false)
-                spinnerSocksPort.setSelection(2)
-            }
-            else -> {
-                editTextSocksPortCustom.setText(initialSocksPort)
-                spinnerSocksPort.setSelection(1)
-                socksPortCustomVisibility(true)
-            }
-        }
-    }
-
-    private fun socksPortCustomVisibility(show: Boolean) {
-        textViewSocksPortCustom.visibility = if (show) View.VISIBLE else View.GONE
-        editTextSocksPortCustom.visibility = if (show) View.VISIBLE else View.GONE
-    }
-
-    private fun saveSocksPort(): Any? {
-        val socksPort = when (adapterSocksPort.getItem(listenerSocksPort.position)) {
-            AUTO -> {
-                PortOption.AUTO
-            }
-            DISABLED -> {
-                PortOption.DISABLED
-            }
-            else -> {
-                editTextSocksPortCustom.text.toString()
-            }
-        }
-
-        try {
-            serviceTorSettings.socksPortSave(socksPort)
-            initialSocksPort = socksPort
-        } catch (e: IllegalArgumentException) {
-            DashboardFragment.showMessage(
-                DashMessage("${DashMessage.EXCEPTION}${e.message}",
-                    R.drawable.dash_message_color_red,
-                    4_000
-                )
-            )
-            return null
-        }
-        return Any()
-    }
-
-    /////////////////
-    /// Http Port ///
-    /////////////////
-
-    private inner class TorOptionSelectionListener: AdapterView.OnItemSelectedListener {
-        var position = 0
-            private set
-        var count = 0
-            private set
-
-        override fun onNothingSelected(parent: AdapterView<*>?) {}
-
-        override fun onItemSelected(parent: AdapterView<*>?, view: View?, position: Int, id: Long) {
-            this.position = position
-
-            if (count == 0) {
-                count ++
-                return
-            }
-
-            val item = parent?.getItemAtPosition(position) ?: return
-
-            when (parent.adapter) {
-                adapterSocksPort -> {
-                    when (item.toString()) {
-                        AUTO -> {
-                            socksPortCustomVisibility(false)
-                        }
-                        DISABLED -> {
-                            socksPortCustomVisibility(false)
-                        }
-                        else -> {
-                            when (initialSocksPort) {
-                                PortOption.AUTO, PortOption.DISABLED -> {
-                                    editTextSocksPortCustom.setText("")
-                                }
-                                else -> {
-                                    editTextSocksPortCustom.setText(initialSocksPort)
-                                }
-                            }
-                            socksPortCustomVisibility(true)
-                        }
-                    }
-                }
-            }
         }
     }
 }
